# internal/handler/v1/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a swagger specification for a Duel Duck back-end.",
        "title": "Duel Duck API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/faq": {
            "get": {
                "description": "Retrieve a list of all FAQs with optional filtering, sorting, and pagination.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "faq"
                ],
                "summary": "Get all FAQs",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Page number (starting from 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of FAQs with pagination info",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing claims",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new FAQ entry with the provided question and optional images.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "faq"
                ],
                "summary": "Create a new FAQ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Question text",
                        "name": "question",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "csv",
                        "description": "Images to upload",
                        "name": "images",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "FAQ created successfully",
                        "schema": {
                            "$ref": "#/definitions/model.FAQ"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing claims",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    }
                }
            }
        },
        "/faq/mark": {
            "put": {
                "description": "Adds a mark to an existing FAQ entry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "faq"
                ],
                "summary": "Add a mark to an FAQ",
                "parameters": [
                    {
                        "description": "Mark request data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AddFAQMarkReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing claims",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apperrors.ErrorPublic": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "model.AddFAQMarkReq": {
            "type": "object",
            "properties": {
                "question_id": {
                    "type": "integer"
                },
                "state": {
                    "type": "integer"
                }
            }
        },
...
        "model.FAQ": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "answer_created_at": {
                    "type": "string"
                },
                "answer_id": {
                    "type": "integer"
                },
                "answer_images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "is_answered": {
                    "type": "boolean"
                },
                "is_marked": {
                    "type": "boolean"
                },
                "is_me": {
                    "type": "boolean"
                },
                "marks_value": {
                    "type": "integer"
                },
                "question": {
                    "type": "string"
                },
                "question_created_at": {
                    "type": "string"
                },
                "question_images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    }
}